requirements.txt
django~=4.0
django-phonenumber-field[phonenumberslite]
python-dotenv
tox
